51 c

+++ C/Medium/51.c
@@ -1,10 +1,10 @@
 
 #include <stdio.h>
-void myFunction(int a);
+void myFunction(float a);
 int main() {
     myFunction(5.5);
     return 0;
 }
-void myFunction(int a) {
-    printf("%d\n", a);
+void myFunction(float a) {
+    printf("%f\n", a);
 }
